pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowlens"
version = "0.1.0"
description = "Application Dependency Mapping platform - ingests network flow data, builds dependency graphs, exposes REST API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "FlowLens Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
]
keywords = ["netflow", "sflow", "ipfix", "dependency-mapping", "network-monitoring"]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "uvloop>=0.19.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # Validation & Settings
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Logging & Observability
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",

    # Networking & Parsing
    "dnspython>=2.4.0",
    "maxminddb>=2.5.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Utilities
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
# High-scale components (only when >10k flows/sec)
kafka = [
    "aiokafka>=0.10.0",
    "confluent-kafka>=2.3.0",
]
redis = [
    "redis[hiredis]>=5.0.0",
]
scale = [
    "flowlens[kafka,redis]",
]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.26.0",
    "respx>=0.20.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "testcontainers[postgres]>=3.7.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

all = [
    "flowlens[scale,dev,docs]",
]

[project.scripts]
flowlens-api = "flowlens.api.main:run"
flowlens-ingestion = "flowlens.ingestion.main:run"
flowlens-enrichment = "flowlens.enrichment.main:run"
flowlens-resolution = "flowlens.resolution.main:run"

[project.urls]
Homepage = "https://github.com/flowlens/flowlens"
Documentation = "https://flowlens.github.io/flowlens"
Repository = "https://github.com/flowlens/flowlens"
Issues = "https://github.com/flowlens/flowlens/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/flowlens"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/migrations",
    "/tests",
]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["flowlens"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG001", "S101"]
"migrations/**/*.py" = ["ERA001"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin",
]

[[tool.mypy.overrides]]
module = [
    "uvloop",
    "asyncpg",
    "maxminddb",
    "aiokafka",
    "confluent_kafka",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:passlib.*",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests requiring external services",
    "slow: Slow tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/flowlens"]
branch = true
parallel = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
fail_under = 80
